pyproject.toml
[project]
name = "my-rc4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
